+++ changelog.md
@@ -2,6 +2,8 @@
 
 * Backwards compatibility with GHC 7.10.
 
+* Fix `Store` instances and test them.
+
 
 # Version 0.1